Refactor to use middleware from weaveworks/common.
