+++ .travis.yml
@@ -3,7 +3,7 @@ sudo: required
 services:
   - docker
 go:
-  - 1.6
+  - 1.7
 before_install:
   - go get -u github.com/FiloSottile/gvt
   - gvt restore
