+++ scripts/build.sh
@@ -27,7 +27,6 @@ echo $CODE_DIR
  
 cp -r $CODE_DIR/images/ $CODE_DIR/docker/catalogue/images/
 cp -r $CODE_DIR/cmd/ $CODE_DIR/docker/catalogue/cmd/
-cp -r $CODE_DIR/middleware/ $CODE_DIR/docker/catalogue/middleware/
 cp $CODE_DIR/*.go $CODE_DIR/docker/catalogue/
 mkdir -p $CODE_DIR/docker/catalogue/vendor && cp $CODE_DIR/vendor/manifest $CODE_DIR/docker/catalogue/vendor/
