+++ test/unit.py
@@ -11,7 +11,7 @@ class GoServices(unittest.TestCase):
         code_dir = script_dir + "/.."
         home = expanduser("~")
         goPath = os.environ['GOPATH']
-        command = ['docker', 'run', '--rm', '-v', goPath + ':/go/src/', '-v', code_dir + ':/go/src/github.com/microservices-demo/catalogue', '-w', '/go/src/github.com/microservices-demo/catalogue', '-e', 'GOPATH=/go/', 'golang:1.6', 'go', 'test', '-v', '-covermode=count', '-coverprofile=coverage.out']
+        command = ['docker', 'run', '--rm', '-v', goPath + ':/go/src/', '-v', code_dir + ':/go/src/github.com/microservices-demo/catalogue', '-w', '/go/src/github.com/microservices-demo/catalogue', '-e', 'GOPATH=/go/', 'golang:1.7', 'go', 'test', '-v', '-covermode=count', '-coverprofile=coverage.out']
 
         print(Docker().execute(command))
